+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.3.1] - 2022-06-19
+
+### Changed
+
+- Improved panic messages produced by `InjectX()` functions.
+
 ## [0.3.0] - 2022-06-19
 
 ### Added
@@ -38,6 +44,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.1.0]: https://github.com/dogmatiq/imbue/releases/tag/v0.1.0
 [0.2.0]: https://github.com/dogmatiq/imbue/releases/tag/v0.2.0
 [0.3.0]: https://github.com/dogmatiq/imbue/releases/tag/v0.3.0
+[0.3.1]: https://github.com/dogmatiq/imbue/releases/tag/v0.3.1
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD